Backend/pyproject.toml
[project]
name = "rag-medical-chatbot"
version = "0.1.0"
description = "Medical assistant chatboot with RaG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-session>=0.8.0",
    "huggingface-hub>=0.35.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-huggingface>=0.3.1",
    "pip>=25.2",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sentence-transformers>=5.1.1",
]
